Link history from Spark 3.5 to 4.0
