--- plugins/spark/v3.5/spark/checkstyle_suppressions.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<!--
- Licensed to the Apache Software Foundation (ASF) under one
- or more contributor license agreements.  See the NOTICE file
- distributed with this work for additional information
- regarding copyright ownership.  The ASF licenses this file
- to you under the Apache License, Version 2.0 (the
- "License"); you may not use this file except in compliance
- with the License.  You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing,
- software distributed under the License is distributed on an
- "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- KIND, either express or implied.  See the License for the
- specific language governing permissions and limitations
- under the License.
--->
-<!DOCTYPE suppressions PUBLIC
-  "-//Puppy Crawl//DTD Suppressions Configuration 1.2//EN"
-  "https://checkstyle.org/dtds/suppressions_1_2.dtd">
-
-<suppressions>
-  <!--
-   Suppress IllegalImport in all files under plugins/spark/v3.5/spark and common.
-   The spark client relies on the shaded libraries from iceberg-spark-runtime,
-   and therefore uses imports like org.apache.iceberg.shaded.* in the code,
-   which is intended.
-  -->
-  <suppress checks="IllegalImport" files="plugins/spark/(v3.5/spark|common)/.*"/>
-</suppressions>
