+++ plugins/spark/v4.0/spark/BUNDLE-LICENSE
@@ -204,8 +204,8 @@
 
 This product bundles and includes code from Apache Iceberg.
 
-* plugins/spark/v4.0/spark/src/main/java/org/apache/polaris/spark/PolarisRESTCatalog.java
-* plugins/spark/v4.0/spark/src/main/java/org/apache/polaris/spark/SparkCatalog.java
+* plugins/spark/v3.5/spark/src/main/java/org/apache/polaris/spark/PolarisRESTCatalog.java
+* plugins/spark/v3.5/spark/src/main/java/org/apache/polaris/spark/SparkCatalog.java
 
 Copyright: 2017-2025 The Apache Software Foundation
 Project URL: https://iceberg.apache.org/
