+++ plugins/spark/v4.0/spark/src/main/java/org/apache/polaris/spark/utils/PolarisCatalogUtils.java
@@ -158,12 +158,11 @@ public class PolarisCatalogUtils {
             Option.apply(genericTable.format()),
             emptyStringSeq,
             scala.Option.empty(),
-            genericTable.properties().getOrDefault(TableCatalog.PROP_OWNER, ""),
+            genericTable.properties().get(TableCatalog.PROP_OWNER),
             System.currentTimeMillis(),
             -1L,
             "",
             scalaOptions,
-            scala.Option.empty(),
             scala.Option.empty(),
             scala.Option.empty(),
             scala.Option.empty(),
